block devices: use 32bit block index

+++ src/kernel/vfs/device_ops.h
@@ -27,10 +27,10 @@ typedef struct block_device_ops block_device_ops_t;
 struct block_device_ops {
     // seems bad: shouuld return 0
     // return the number of block read on sucess (1)
-    int (*read_block)(void *private, uint64_t index, void *block);
+    int (*read_block)(void *private, uint32_t index, void *block);
 
     // return the number of block written if success (1)
-    int (*write_block)(void *private, uint64_t index, const void *block);
+    int (*write_block)(void *private, uint32_t index, const void *block);
 };
 
 typedef struct block_device {
